Fix load

+++ models/base_model.py
@@ -177,7 +177,8 @@ class BaseModel(ABC):
                     save_paths.append(os.path.join("/content/gdrive/My Drive/", self.opt.name, save_filename))
 
                 for save_path in save_paths:
-                    os.remove(save_path)
+                    if os.path.exists(save_path):
+                        os.remove(save_path)
 
 
     def __patch_instance_norm_state_dict(self, state_dict, module, keys, i=0):
@@ -205,8 +206,8 @@ class BaseModel(ABC):
                 load_filename = '%s_net_%s.pth' % (epoch, name)
                 load_path = os.path.join(self.save_dir, load_filename)
                 net = getattr(self, 'net' + name)
-                if isinstance(net, torch.nn.DataParallel):
-                    net = net.module
+                #if isinstance(net, torch.nn.DataParallel):
+                    #net = net.module
                 print('loading the model from %s' % load_path)
                 # if you are using PyTorch newer than 0.4 (e.g., built from
                 # GitHub source), you can remove str() on self.device
@@ -216,8 +217,8 @@ class BaseModel(ABC):
 
                 # patch InstanceNorm checkpoints prior to 0.4
                 for key in list(state_dict.keys()):  # need to copy keys here because we mutate in loop
-                    self.__patch_instance_norm_state_dict(state_dict, net, key.split('.'))
-                net.load_state_dict(state_dict)
+                    self.__patch_instance_norm_state_dict(state_dict, net.module, key.split('.'))
+                net.module.load_state_dict(state_dict)
 
     def print_networks(self, verbose):
         """Print the total number of parameters in the network and (if verbose) network architecture
